Add PEP503 norm to new/upd package checking
If a package gets renamed to within identical PEP503 normalization,
the version tracking should be able to follow this rename.

This rev implements this, as confirmed by flufl.flake8 being correctly
treated as an update to flufl-flake8.

Curiously, both flufl.flake8 and flufl-flake8 are appearing in
the rendered Markdown tables. Seems like flufl-flake8 should now
not be showing up there...